Patch from Laurent.  In his words:
I'm proposing to add a new exception (JDOMParseException), subclass of
JDOMException, to be thrown by the builders to convey whatever information
could be gathered from the parser about the error.
The implementation with SAXBuilder is trivial and does not require code
changes in the existing applications.

The really interesting part here is the getPartialDocument method of
JDOMParseException that gives access to whatever part of the input document
that was successfully parsed before the parser fired a SAXParseException.

-jh-