New feature from Laurent.  In his words:
--
While trying fix a problem in a class that feeds the output of SAXOutputter
into another tool, I found it would be helpful if SAXOutputter could provide
the information of which node was being processed at the time the error occurred.
To stay in the SAX "philosophy", the cleanest way was to use a Locator object
(JDOMLocator) that exposes the node being processed.
--

Note this adds a new public class.  If anyone has problems with it, write
in.

-jh-