Patch from Laurent.  In his words:
I have enhanced SAXOutputter regarding the way JDOMLocator is managed.

The diff file lists theses changes compared to the version you have committed:

  - Reset the locator once the output is done. As this is not a crucial
operation, I did not include it in a try/finally block but it's better to
clean it when possible.

  - Add a new getLocator method to make the locator available outside the
ContentHandler. For example, allows to have ErrorHandlers not implementing
XMLFilters.

-jh-