Patch from Rusty:
Here's a patch for the TestDocument class that removes a \r\n from the
expected output of the toString() method in Document so that the test
passes now. The return value of toString() shouldn't contain a line
break in any case. The Document class is right. The source code is
wrong.

-jh-