From Rolf Lear:
The following provides a slight, but easy performance gain in
Verifier.checkXMLName().

Existing code:
...
for (int i=0, len = name.length(); i<len; i++) {
...


Replace with:
for (int i=1, len = name.length(); i<len; i++) {

Reason:
The first character is already confirmed to be an "isXMLNameStartCharacter",
which is a pure subset of the "isXMLNameCharacter" every other letter is
checked against.  Thus, the first letter of every Element and Attribute name
is verified twice.

-jh-