Greatly improving and standardizing the class-level Javadocs.  Method-level documentation hasn't been touched.
The @version is coming ahead of @author which isn't "standard" per

  http://java.sun.com/j2se/javadoc/writingdoccomments/

but it's how the JavadocFormatter (JDF) plug-in in IntelliJ does things.
I figure if/when they fix that, we'll get things all ordered back again.

Since we're looking so much better I'm tempted to change the web site
Javadocs.

-jh-