Changed the Document's DocType storage so it's part of the ContentList now and has a location that can be preserved.  This helps with round tripping.
The getDocType() and setDocType() methods remain for convenience but
just operate based on searches through the ContentList.

Deprecated the Document(List, DocType) constructor because it doesn't
make sense if the DocType should be part of the content list rather than
separate.

Added Document.addContent(DocType) which appends the DocType to the
ContentList.  setDocType() does an insert at the head which is what you
normally would want, but builders need to use addContent() so it's an
append into the right location.

I think I got all the parentage checks and manipulations correct, but
it's gotten extremely complicated and dispersed.

Removed this item from the TODO.

-jh-