Added an EscapeStrategy plug-in interface for XMLOutputter to determine which chars to escape.  A user can set a strategy and go, no need to subclass.
Also created a DefaultEscapeStrategy which tries to be generally smart.
It quickly says no escaping is necessary for UTF-8 (our default) and UTF-16.
It escapes everything > 255 for ISO-8859-1/Latin1.  It escapes everything
> 127 for ASCII.

For the other charsets it tries to use the JDK 1.4 CharsetEncoder to
determine if the char needs escaping.  I use reflection for this so
we aren't dependent on JDK 1.4  That means if you run on JDK 1.3 there's
no escaping unless we know about the charset ourselves.

I'm not sure how slow the repeated reflection invoke() is.  I'm guessing
it's pretty heavy, but I don't know how else to tackle that problem and
not depend on 1.4.  People can always plug-in their own strategy if they
want.

Thanks for Alex and Brad for the ideas behind this.  Please guys, double
check my code and if someone wants to do a before/after perf test that'd
be great!

-jh-