Adding the getAttributeType() item to the TODO per Rusty's suggestion:
--
>I'm going to punt on this right now because I'm actually looking to
>bring in Brad's jdom-x Parent/Child interfaces which may or may not
>involve a getType() style method.


OK, but can we add a note to the TODO to consider this again after a
final decision has been made on Parent/Child so I can delete this one
from my Inbox? I understand why you want to delay resolution, but if
we don't go with Parent/Child this is still a concern (and maybe if
we do).
--

-jh-