Patch from Laurent:
ElementScanner notifies the listeners before forwarding the endElement
event to SAXHandler. If a listener detaches the element from its parent,
SAXHandler sets the being-built element to null (the parent), hence the
NPE reported by Xerces.

Another side effect of this bug was that the last text segment of the
element content was missing from the notified elements.

The attached patch fixes the problem.

-jh-