BIG CHANGE.  Comments welcome!
Added two new interfaces, Parent and Child.

Parent has methods (* means new):
  * int getChildCount();
    Parent addContent(Child child);
  * Parent addContent(Collection collection);
  * Parent addContent(int index, Child child);
  * Parent addContent(int index, Collection collection);
  * List cloneContent();
    List getContent();
    List getContent(Filter filter);
  * Child getContent(int index);
  * List removeContent();
  * List removeContent(Filter filter);
  * Child removeContent(int index);
    boolean removeContent(Child child);
    Parent setContent(Child child);
    Parent setContent(Collection collection);
  * Parent setContent(int index, Child child);
  * Parent setContent(int index, Collection collection);
    Object clone();

Child has methods:
    Child detach();
    Document getDocument();
    Parent getParent();
  * String getValue();
    Object clone();

Parent is implemented by Document and Element.

Child is implemented by Comment, DocType, Element, EntityRef,
ProcessingInstruction, and Text (CDATA).

The new methods on Parent are pretty self explanatory.  A few methods that
used to require List work are now on Parent itself for convenience.  The
cloneContent() and removeContent() calls should be especially helpful.

The getValue() method in Child is defined to return the XPath 1.0 string value
of the element.  The getText() methods in Element are left unchanged.

A subtle change is that getParent() now returns a Parent type which is its
immediate parent.  Previously an item at the document level would return null
and you'd use getDocument() to get its Document.

In a similar vein, the protected setDocument() methods have been removed in
favor of just using setParent().  getDocument() remains as a potentially
recursive lookup method.

I renamed getChildren() to getChildElements() and renamed getChild() to
getChildElement().  Now that there's a real Child interface getChildren() is
even more confusing than it used to be.  I deprecated the old forms.  This
change will require a lot of recoding for people.

Also updated supporting files to use the new APIs.

Thanks to Brad's jdom-x experimental tree for lots of the ideas in here,
and thanks to Brad for pre-reviewing this change.

-jh-