Added an UncheckedJDOMFactory class which builds without doing any content or structure checks, letting you gain speed in a situation where you have 100% confidence in your parser.  The Javadocs for the class naturally includes a serious warning.
It's not used by default, but you can select it with a
builder.setFactory(new UncheckedJDOMFactory()) call.

I also added to JDOMFactory a few methods:
  addContent(Parent, Content)
  setAttribute(Element, Attribute)
  addNamespaceDeclaration(Element, Namespace)

These are called during the build to do the adds.  The default builder
just calls parent.addContent(Content) while the "unchecked" factory does
the work without checks using package protected methods on ContentList
and AttributeList.

A perk of having these methods in the factory and used by the builder
is you can write a custom factory to do certain things during the adds.
Like you could ignore all elements named "foo" by not doing the add if
the Content was an Elt foo.  That's not perfect since the elements
underneath foo would still be built into a subtree that got ignored,
but it's an easy solution to save memory in the resulting document.

Tagged the distribution as BEFORE_UNCHECKED since this is a significant
change we might want to back out.

Thanks to Philippe Riand at Trilog for starter code I used to build this.
His code (mailed to the list) required several new classes, and I changed
that to the pluggable factory model.

I'll ask Phil to test this change using his Win32 library that gets CPU
time.

-jh-