Patch from Edwin Wiles (aka "spindizzy").  From his email:
--
This was a problem where DOMOutputter was adding a namespace
declaration to an element that used an attribute from a
different namespace, even though the attribute's namespace
was already "in scope".

Original XML Document:

<?xml version="1.0" encoding="UTF-8"?>
<test xmlns="http://foo/test#" xmlns:rdf="http://pseudo/rdf">
	<rdf:embedded rdf:ID="1">
		<rdf:type rdf:resource="sometypehere" />
		<otherobject rdf:resource="resourcerefhere" />
		<otherprop>propvalue</otherprop>
	</rdf:embedded>
</test>

Result of Text -> JDOM -> W3C DOM -> Text (note the extra namespace):

<?xml version="1.0" encoding="UTF-8"?>
<test xmlns="http://foo/test#" xmlns:rdf="http://pseudo/rdf">
    <rdf:embedded rdf:ID="1">
        <rdf:type rdf:resource="sometypehere"/>
        <otherobject rdf:resource="resourcerefhere" xmlns:rdf="http://pseudo/rdf"/>
        <otherprop>propvalue</otherprop>
    </rdf:embedded>
</test>

Edwin tested this patch and it removes the extra namespace.

-jh-