Patch developed by Brad and me based on a bug report from Wolfgang Frech:
--
FilterIterator does not comply to the java.util.Iterator spec in the
following detail:

If next() is called on a non-empty FilterIterator without a preceding
hasNext(), the FilterIterator throws a NoSuchElementException.
The expected behavior is: next returns the next element, regardless of
preceding calls to hasNext.

This problem can be avoided by a straightforward workaround: always call
hasNext() before next().

A comment at FilterIterator.remove hints that the author is aware of a
similar problem with a call to remove before hasNext.
--

Simple one-liner.

-jh-