First, added the factory method entityRef(String elementName, String systemID) to match the EntityRef constructor with the same parameter signature.  We've been matching constructors with factory calls but overlooked this one.
People who wrote their own factories and didn't subclass
DefaultJDOMFactory will need to take note!

Second, moved addContent() and setContent() from Parent into Document
and Element directly.  This restores the method chaining that some
people really like.  It behaves more like Beta 9 code now.  This could break code that changed for b10 if it called the methods on the Parent
interface.

When we move to J2SE 5.0 we can use covariant return types to put the
methods in Parent while avoiding the cast in Document and Element.
That's the ideal.

-jh-