Dietrich Schulten reported:
--
for some reason, the Ids contribution does not work at the moment.The
ID hashmap is always empty. In the method IdElement.setAttribute(),
where the ID should be added, this.getDocument() is always null.
--

This is a patch from Laurent to fix the issue.  Dietrich hasn't verified
the fix worked.  He fell silent.  I'll just trust Laurent.

Note from Laurent:

--
This version adds support for updating the IdDocument lookup table when
modifying the ID attribute value or type.
--

-jh-