Changed signature of
 public Parent setContent(int index, Collection collection)

to

 public Element setContent(int index, Collection collection)

Because all the other setContent() and addContent() methods
return Element.  The Parent was a bug oversight.

I'm not real pleased with changing this in 1.0.1 but I think it's
the right thing.  It's a little used method and the return value
is even more little used.  Plus people who used the return value
would work around the Parent return value by casting the return
value to Element which will continue to work after this change.
Convenient!  The downside is that people who use the method may
need to recompile to work with the new signature since Java won't
see them as identical method signatures.

-jh-