Committed a patch set from Victor Toni.
He fixed the bug he reported earlier where the attribute type would
change on a setAttribute() call.

He did a stylistic fix to remove the cast to Element in the loop of
the isAncestor() method which was unnecessary because one can iterate
over the returned parent.

He changed almost all local variables and parameters to final.
This doesn't affect compiled code or inheritance.  It's just an
optimization and a safety check.

-jh-