Fix from Victor Toni:
While working on the test cases for CDATA I looked up the code and
noticed that a corner case is not handled correctly in particular
append() could lead to invalid content:
Text set is checked if it contains "]]>", when appending  text only the
new part is checked and not the resulting value leading to this to
corner case which are currently ignored:
"some text ]" + "]> some other text"
and
"some text ]]" + "> some other text"

Comment doesn't seem to have a similar issue since it isn't derived from
Text and thus doesn't have an append method.

Attached you'll find a complete test class for CDATA with a 100%
coverage and the changed CDATA class which solves the issue.
I had to add an override for Text.append(Text) in CDATA since the same
could happen there.

-jh-