Patch from Dave Byrne:
When transforming a Document using XSLTransformer that required an
EntityResolver to be parsed, SAXOutputter.dtdEvents() throws an
exception as it is unable to locate any dtds/entities.

Attached is a patch that adds an additional constructor to JDOMSource
with an EntityResolver which is passed to the internal DocumentReader
allowing the SAXOutputter to properly resolve dtds.

-jh-