Fixed bug reported at http://www.junlu.com/msg/166290.html: --- I got an exception when I executed code along the lines of the following:
  Namespace ns1 = Namespace.getNamespace("aaa", "http://acme.com/aaa");
  Element e = new Element("aaa", ns1);
  e.setAttribute("att1", "att1");
  Namespace defns = Namespace.getNamespace("http://acme.com/default");
  e.addNamespaceDeclaration(defns);

I want to declare an element that is in the 'aaa' namespace but that also
declares a default namespace that is inherited by other elements. However, I
get this exception:

org.jdom.IllegalAddException Source code of org.jdom.IllegalAddException: The
namespace xmlns="http://acme.com/default"
could not be added as a namespace to "aaa:aaa": The namespace prefix ""
collides with an attribute namespace prefix on the element

I think that the code in the Verifier is incorrect. It seems to assume that
attributes that have no explicit namespace are in the default namespace. This
is not my understanding of the XML spec.
---

Reported again today by Jeff Dalton.  Just missed the JDOM 1.1 cutoff.

-jh-