Complete the Verifier test code coverage
The TestVerifier class was re-organized, and a lot of dead code removed.
Further, the method names were renamed, etc.

Of interest here is the new test harness TestVerifierCharacters. This is
basically a generated file, where the tests are built by the code in the
VerifierTestBuilder class. This guarantees 100% code-point coverage for
the Verifier character tests. The assumption is that the Verifier is
currently accurate. This is now useful for regression testing.

The volume of tests in the TestVerifierCharacters class is now large,
and it is slow for some character types. This slows down the coverage
testing. The tests report the time in the test.