Fixes #15 by re-working how ContentList deals with iterators and Filters. The Re-Work includes a major performance upgrade because FilterList and FilterListIterator are now 'lazy' in the sense that they only check as much data as needed to satisfy the user request. The previous version would scan the entire ContentList at constructor time.
This should make things like for-each type loops much faster to start
up, and isEmpty() and random-access-type get() calls much faster too.

Add a whole lot of tests for the
ConcurrentModificationException (CME) concept. Note, CME is
'controversial' because set(), remove(), add(), next() and previous()
methods of the core java.util collection Iterators throw CME, but the
hasNext(), nextIndex(), hasPrevious() and previousIndex() do not.

I believe that the 'right' thing is to fail fast, and if you know that
there is a CME problem in the hasNext() or whatever, then throw the CME
immediately. But, since Java has not been 'fixed', and other 'big' fish
have conformed to the actual Java way of doing things (GCJ, Apache
Harmony, etc.), then who am I to disagree.

So, match the base Java way of doing things, and only throw CME in the
iterator 'change' methods, not just the 'query' methods.

Now that FilterList conforms to java.util.List API, we can add a bunch
of test cases for them.

Now that CME conforms to Java core, we can add test cases for that too.