Major refactor of XMLOutputter. Some deviations from JDOM1: 1. Document output will only ever have one end-of-line before the end-of-file (JDOM1 sometimes put two end-of-line). I consider this to be 'minor'. 2. Although the 'public' API side of XMLOutputter is unchanged, the subclassing and customisation process is very, very different. 3. while XMLOutputter.setFormat() still clones the input format, the getFormat() no longer does, so you can now change the XMLOututter's Format instance without having to create a new instance. XMLOutputter is now a light-weight class, anyway, so is very cheap to create instances of.
This is essentially a re-write of the process, and, based on the jUnit
tests, and some additional tests, I have a high level of confidence in
the result, but there is still the potential of some problem coming
through. It does offer a performance improvement over the JDOM1, but,
more importantly, it introduces a much much cleaner code base, and the
logic for outputting XML is much clearer.

Further, although people who customise the XMLOutputter are rare, it
should now be much simpler to make changes to the process.

The JavaDoc has been completely revised, and updated.